+++ src/extension.js
@@ -1,12 +1,19 @@
 // @ts-nocheck
-// @ts-nocheck
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 'use strict';
 
 const vscode = require('vscode');
-const EGE = require('./ege')
+const cp = require('child_process');
+const os = require('os');
+const fs = require('fs-extra');
 
+const EGE = require('./ege');
+const { stderr } = require('process');
+
+/**
+ * @type {EGE}
+ */
 let egeHandle = null;
 
 /**
@@ -26,16 +33,36 @@ function activate(context) {
 
 	context.subscriptions.push(vscode.commands.registerCommand('ege.setup-global', () => {
 		vscode.window.showInformationMessage("ege: setup-global!!\n");
-
 		egeHandle.performInstall();
+	}));
 
-		// if (!progressHandle) {
+	context.subscriptions.push(vscode.commands.registerCommand('ege.cleanup-caches', () => {
+		egeHandle.clearPluginCache();
+	}));
 
-		// 	progressHandle.start("ege: setup-global!!\n", "ege");
-		// } else {
-		// 	progressHandle.cancel();
-		// 	progressHandle = null;
-		// }
+	context.subscriptions.push(vscode.commands.registerCommand('ege.open-cache-dir', () => {
+		if (egeHandle.egeInstallerDir && fs.existsSync(egeHandle.egeInstallerDir)) {
+			const osName = os.platform();
+			let openExplorerCommand = null;
+			if (osName === 'win32' || osName === 'cygwin') { /// win
+				openExplorerCommand = `explorer.exe "${egeHandle.egeInstallerDir}"`;
+			} else if (osName === 'darwin') { /// mac
+				openExplorerCommand = `open "${egeHandle.egeInstallerDir}"`;
+			} else if (osName === 'linux') {
+				openExplorerCommand = `xdg-open "${egeHandle.egeInstallerDir}"`;
+			} else {
+				vscode.window.showErrorMessage(`ege: open cache dir is not supported on ${egeHandle.egeInstallerDir}`);
+			}
+
+			if (openExplorerCommand) {
+				cp.exec(openExplorerCommand, (err) => {
+					if (err)
+						vscode.window.showErrorMessage(`ege: open cache dir ${egeHandle.egeInstallerDir} failed: ${err}`)
+				});
+			}
+		} else {
+			vscode.window.showErrorMessage(`ege: Cache dir ${egeHandle.egeInstallerDir} does not exist.`)
+		}
 	}));
 }